Add abstract incoming/outgoing edges interface
